pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tg-time-logger"
version = "0.1.0"
description = "Telegram time tracker with fun points economy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
